refactor(agents): extract parallel tool tests into dedicated module
Split the 3 parallel tool execution tests (196 lines) from basic.rs into
parallel.rs to keep basic.rs well under the 1500-line CI limit (1302).

Entire-Checkpoint: 299716696618
